Add RobotController architecture
Update class loader to point to the new physcial robot class

We want to keep the robot's controller separated from the physical robot
for testing purposes. The physical robot defers control to an
IRobotController, and an IRobotController performs actions on an IRobot,
doing things like setting motor speeds or getting access to sensors.