Handle pneumatics in a solenoid-agnostic fashion
The practice robot uses a single solenoid and the competition robot uses
a double solenoid. The single solenoid has no "neutral position," the on
mode fires the solenoid and the off mode returns to the inital position
with a spring. The double solenoid actually has an explicit neutral
mode. This commit adds AbstractSolenoid so that the OneSolenoid and
TwoSolenoid can handle each piece of hardware in the appropriate
fashion.

Swap the TankInput buttons