Add insertNext() to AutoState
In order to run the LIDAR sensor autonomous, we need to be able to
dynamically add new states inbetween the currently existing states.
insertNext() will take a chain like A -> C and make it A -> B -> C,
whereas setNext() would take A -> C and make it A -> C -> B.

Add tests.