Move state dependent code into AutoState init()
Previously, creating an AutoState would set some state that needed
to be correct when the robot began that state. For instance,
ForwardAutonomous would set the initial time when the state was
created, rather then when the state began, which meant that the time
was usually up by the time it started. Setting the initial time
during init() fixes this problem.