Make DistanceEncoder return correct distances.
DistanceEncoder used to multiply all the constants together when returning values, which was wrong. This commit fixes that.

The fuctions which handle this have been refactored to be clearer.

In addition, it is now possible to convert from inches back to counts.

Also included are a lot of code cleanups.