Add the LocateTargetAutonomous
The LocateAutonomous will begin turning in a direction attempting to see the
vision target and then use that information in order to finish the turn
accurately.

Add vision staleness
Using a vision value that isn't recent enough is often worse than useless.
If you think you need to turn left and you actually need to turn right, things
will keep getting worse. Here we add an API to find out if the vision heading
is recent enough, so that we can avoid acting on stale data.